se creó vista/ con MenuInicial.java, TableroJuego.java y CartaPanel.java; se creó controlador/ControladorJuego.java para centralizar la lógica de juego; se mantuvo modelo/ con las clases existentes; se refactorizó TableroJuego para que sea solo UI; se movió la lógica de juego a ControladorJuego; se actualizó MenuInicial para iniciar el controlador y abrir la vista; se actualizó App.java para lanzar MenuInicial; y se eliminó classes/ para dejar solo el código fuente en src
