+++ yugi_oh_JAVA/src/vista/CartaPanel.java
@@ -1,3 +1,5 @@
+package vista;
+
 import javax.swing.*;
 
 import modelo.Carta;
